Move SupabaseProvider to lib
