+++ talentify-next-frontend/app/talent/dashboard/layout.tsx
@@ -6,7 +6,7 @@ import Sidebar from "@/components/Sidebar"
 import Footer from "@/components/Footer"
 import { Inter, Noto_Sans_JP } from "next/font/google"
 import { createClient } from "@/lib/supabase/server"
-import { SupabaseProvider } from "@/utils/supabase/provider"
+import { SupabaseProvider } from "@/lib/supabase/provider"
 
 const inter = Inter({
   subsets: ["latin"],
